fun rewrite time
